Makefile: use <project>_DATEDEPS variables for dates
List the projects used to calculate date in the variables instead of
directly in the rules. It helps to remove hptzip condition in hpt
and htick rules and will help later to make a universal rule.

For since need_hptzip variable is 0/1, use another one,
HPTZIP (blank or hptzip) to use in the lists. The list variables
must be lazy for that reason.

Keep hpt and htick *_cmp indention to make the patch more readable.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>