Makefile: replace git update with a static pattern rule
The rule is the same for all the update targets, replace it with the
pattern rule to avoid duplication.

It can be possibly better to generate it instead when the rules will
be generated.

The rule must be static pattern rule since it's .PHONY and implicit
pattern rules is not rebuilt when the dependency file does not
exist.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>