Makefile: cvsdate.h: generate with gen_cvsdate function
Unify cvsdate.h creation and move it to a function.
There are 2 types of it:
- using make variable when date is taken from the subproject only;
- using shell variable when date is selected from several
  subprojects.

Use only the latter one. To do it the rules which used the former
way need to initialize shell variables with the make ones. Do it
with a helper function (it is easier to replace).

For all the subprojects except hptsqfix the file is created in the
root directory, but in hptsqfix it's in the headers' one. So, make
it possible to list exceptions in a variable <project>.CVSDATEDIR.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>