Implement gen_date_selection function
and use it in *_cmp targets to generate shell code selecting the
latest date from dependent subprojects.

List the projects used to calculate date in the variables instead of
directly in the rules. It helps to remove hptzip condition in hpt
and htick rules and will help later to make a universal rule.

Since need_hptzip variable is 0/1, take another one,
HPTZIP (blank or hptzip) to use in the lists. The list variables
must be lazy for that reason.

Keep hpt and htick *_cmp indention to make the patch more readable.

Thanks to Zhenja Kaliuta 2:4500/1.59