move per-subproject generation right before rules
The idea is to have per-subproject variables to be generated from
the function. It must follow special per-subproject variables
section.

Since it also generates _update rules, the first of
them becomes the default, so set .DEFAULT_GOAL. Setting MAKECMDGOALS
affects the code but does not affect default goal selection.

Thanks to Zhenja Kaliuta 2:4500/1.59